Refactor unit creation, implement move action and update actions in loop, better zoom

+++ src/timer.c
@@ -15,12 +15,12 @@ struct timer timer_init() {
   return t;
 }
 
-double timer_elapsed_ms(struct timer *timer) {
+float timer_elapsed_ms(struct timer *timer) {
   struct timeval tv_now;
 
   gettimeofday(&tv_now, NULL);
-  double ms = (tv_now.tv_sec - timer->tv_prev.tv_sec) * 1000.0f +
-              (tv_now.tv_usec - timer->tv_prev.tv_usec) / 1000.0f;
+  float ms = (tv_now.tv_sec - timer->tv_prev.tv_sec) * 1000.0f +
+             (tv_now.tv_usec - timer->tv_prev.tv_usec) / 1000.0f;
   timer->tv_prev = tv_now;
 
   return ms;
